+++ src/extensions/opentelemetry.rs
@@ -1,9 +1,10 @@
 use std::collections::HashMap;
+use std::sync::Arc;
 
 use async_graphql_parser::types::ExecutableDocument;
 use async_graphql_value::Variables;
 use opentelemetry::trace::{SpanKind, TraceContextExt, Tracer};
-use opentelemetry::{Context as OpenTelemetryContext, Key, KeyValue};
+use opentelemetry::{Context as OpenTelemetryContext, Key};
 
 use crate::extensions::{Extension, ExtensionContext, ExtensionFactory, ResolveInfo};
 use crate::{ServerError, ValidationResult};
@@ -27,42 +28,41 @@ const KEY_ERROR: Key = Key::from_static_str("graphql.error");
 const KEY_COMPLEXITY: Key = Key::from_static_str("graphql.complexity");
 const KEY_DEPTH: Key = Key::from_static_str("graphql.depth");
 
-/// OpenTelemetry extension
+/// OpenTelemetry extension configuration for each request.
 #[derive(Default)]
 #[cfg_attr(docsrs, doc(cfg(feature = "opentelemetry")))]
-pub struct OpenTelemetry<T, U> {
-    tracer: T,
-    #[allow(dead_code)]
-    uninstall: U,
+pub struct OpenTelemetryConfig {
+    /// Use a context as the parent node of the entire query.
+    parent: spin::Mutex<Option<OpenTelemetryContext>>,
 }
 
-impl<T> OpenTelemetry<T, ()> {
+impl OpenTelemetryConfig {
+    /// Use a context as the parent of the entire query.
+    pub fn parent_context(mut self, cx: OpenTelemetryContext) -> Self {
+        *self.parent.get_mut() = Some(cx);
+        self
+    }
+}
+
+/// OpenTelemetry extension
+#[cfg_attr(docsrs, doc(cfg(feature = "opentelemetry")))]
+pub struct OpenTelemetry<T> {
+    tracer: Arc<T>,
+}
+
+impl<T> OpenTelemetry<T> {
     /// Use `tracer` to create an OpenTelemetry extension.
-    pub fn new(tracer: T) -> OpenTelemetry<T, ()>
+    pub fn new(tracer: T) -> OpenTelemetry<T>
     where
-        T: Tracer + Send + Sync + Clone,
+        T: Tracer + Send + Sync,
     {
         Self {
-            tracer,
-            uninstall: (),
-        }
-    }
-
-    /// Attach an Uninstall instance to this extension.
-    pub fn with_uninstall<U>(self, uninstall: U) -> OpenTelemetry<T, U>
-    where
-        U: Send + Sync + 'static,
-    {
-        OpenTelemetry {
-            tracer: self.tracer,
-            uninstall,
+            tracer: Arc::new(tracer),
         }
     }
 }
 
-impl<T: Tracer + Send + Sync + Clone, U: Send + Sync + 'static> ExtensionFactory
-    for OpenTelemetry<T, U>
-{
+impl<T: Tracer + Send + Sync> ExtensionFactory for OpenTelemetry<T> {
     fn create(&self) -> Box<dyn Extension> {
         Box::new(OpenTelemetryExtension {
             tracer: self.tracer.clone(),
@@ -72,7 +72,7 @@ impl<T: Tracer + Send + Sync + Clone, U: Send + Sync + 'static> ExtensionFactory
 }
 
 struct OpenTelemetryExtension<T> {
-    tracer: T,
+    tracer: Arc<T>,
     contexts: HashMap<usize, OpenTelemetryContext>,
 }
 
@@ -92,36 +92,46 @@ impl<T> OpenTelemetryExtension<T> {
     }
 }
 
-impl<T: Tracer + Send + Sync + Clone> Extension for OpenTelemetryExtension<T> {
+impl<T: Tracer + Send + Sync> Extension for OpenTelemetryExtension<T> {
+    fn start(&mut self, ctx: &ExtensionContext<'_>) {
+        let request_cx = ctx
+            .data_opt::<OpenTelemetryConfig>()
+            .and_then(|cfg| cfg.parent.lock().take())
+            .unwrap_or_else(|| {
+                OpenTelemetryContext::current_with_span(
+                    self.tracer
+                        .span_builder("request")
+                        .with_kind(SpanKind::Server)
+                        .start(&*self.tracer),
+                )
+            });
+        self.enter_context(REQUEST_CTX, request_cx);
+    }
+
+    fn end(&mut self, _ctx: &ExtensionContext<'_>) {
+        self.exit_context(REQUEST_CTX);
+    }
+
     fn parse_start(
         &mut self,
         _ctx: &ExtensionContext<'_>,
         query_source: &str,
         variables: &Variables,
     ) {
-        let request_span = self
-            .tracer
-            .span_builder("request")
-            .with_kind(SpanKind::Server)
-            .start(&self.tracer);
-        let request_cx = OpenTelemetryContext::current_with_span(request_span);
-        self.enter_context(REQUEST_CTX, request_cx.clone());
-
-        let mut attributes = Vec::with_capacity(2);
-        attributes.push(KeyValue::new(KEY_SOURCE, query_source.to_string()));
-        attributes.push(KeyValue::new(
-            KEY_VARIABLES,
-            serde_json::to_string(variables).unwrap(),
-        ));
-        let parse_span = self
-            .tracer
-            .span_builder("parse")
-            .with_kind(SpanKind::Server)
-            .with_attributes(attributes)
-            .with_parent_context(request_cx)
-            .start(&self.tracer);
-        let parse_cx = OpenTelemetryContext::current_with_span(parse_span);
-        self.enter_context(PARSE_CTX, parse_cx);
+        if let Some(parent_cx) = self.contexts.get(&REQUEST_CTX).cloned() {
+            let mut attributes = Vec::with_capacity(2);
+            attributes.push(KEY_SOURCE.string(query_source.to_string()));
+            attributes.push(KEY_VARIABLES.string(serde_json::to_string(variables).unwrap()));
+            let parse_span = self
+                .tracer
+                .span_builder("parse")
+                .with_kind(SpanKind::Server)
+                .with_attributes(attributes)
+                .with_parent_context(parent_cx)
+                .start(&*self.tracer);
+            let parse_cx = OpenTelemetryContext::current_with_span(parse_span);
+            self.enter_context(PARSE_CTX, parse_cx);
+        }
     }
 
     fn parse_end(&mut self, _ctx: &ExtensionContext<'_>, _document: &ExecutableDocument) {
@@ -135,7 +145,7 @@ impl<T: Tracer + Send + Sync + Clone> Extension for OpenTelemetryExtension<T> {
                 .span_builder("validation")
                 .with_kind(SpanKind::Server)
                 .with_parent_context(parent_cx)
-                .start(&self.tracer);
+                .start(&*self.tracer);
             let validation_cx = OpenTelemetryContext::current_with_span(span);
             self.enter_context(VALIDATION_CTX, validation_cx);
         }
@@ -144,8 +154,8 @@ impl<T: Tracer + Send + Sync + Clone> Extension for OpenTelemetryExtension<T> {
     fn validation_end(&mut self, _ctx: &ExtensionContext<'_>, result: &ValidationResult) {
         if let Some(validation_cx) = self.exit_context(VALIDATION_CTX) {
             let span = validation_cx.span();
-            span.set_attribute(KeyValue::new(KEY_COMPLEXITY, result.complexity as i64));
-            span.set_attribute(KeyValue::new(KEY_DEPTH, result.depth as i64));
+            span.set_attribute(KEY_COMPLEXITY.i64(result.complexity as i64));
+            span.set_attribute(KEY_DEPTH.i64(result.depth as i64));
         }
     }
 
@@ -156,12 +166,12 @@ impl<T: Tracer + Send + Sync + Clone> Extension for OpenTelemetryExtension<T> {
                 .span_builder("execute")
                 .with_kind(SpanKind::Server)
                 .with_parent_context(parent_cx)
-                .start(&self.tracer),
+                .start(&*self.tracer),
             None => self
                 .tracer
                 .span_builder("execute")
                 .with_kind(SpanKind::Server)
-                .start(&self.tracer),
+                .start(&*self.tracer),
         };
         let execute_cx = OpenTelemetryContext::current_with_span(span);
         self.enter_context(EXECUTE_CTX, execute_cx);
@@ -169,7 +179,6 @@ impl<T: Tracer + Send + Sync + Clone> Extension for OpenTelemetryExtension<T> {
 
     fn execution_end(&mut self, _ctx: &ExtensionContext<'_>) {
         self.exit_context(EXECUTE_CTX);
-        self.exit_context(REQUEST_CTX);
     }
 
     fn resolve_start(&mut self, _ctx: &ExtensionContext<'_>, info: &ResolveInfo<'_>) {
@@ -181,19 +190,16 @@ impl<T: Tracer + Send + Sync + Clone> Extension for OpenTelemetryExtension<T> {
 
         if let Some(parent_cx) = parent_cx {
             let mut attributes = Vec::with_capacity(3);
-            attributes.push(KeyValue::new(
-                KEY_RESOLVE_ID,
-                info.resolve_id.current as i64,
-            ));
-            attributes.push(KeyValue::new(KEY_PARENT_TYPE, info.parent_type.to_string()));
-            attributes.push(KeyValue::new(KEY_RETURN_TYPE, info.return_type.to_string()));
+            attributes.push(KEY_RESOLVE_ID.i64(info.resolve_id.current as i64));
+            attributes.push(KEY_PARENT_TYPE.string(info.parent_type.to_string()));
+            attributes.push(KEY_RETURN_TYPE.string(info.return_type.to_string()));
             let span = self
                 .tracer
                 .span_builder(&info.path_node.to_string())
                 .with_kind(SpanKind::Server)
                 .with_parent_context(parent_cx)
                 .with_attributes(attributes)
-                .start(&self.tracer);
+                .start(&*self.tracer);
             let resolve_cx = OpenTelemetryContext::current_with_span(span);
             self.enter_context(resolve_ctx_id(info.resolve_id.current), resolve_cx);
         }
@@ -205,10 +211,9 @@ impl<T: Tracer + Send + Sync + Clone> Extension for OpenTelemetryExtension<T> {
 
     fn error(&mut self, _ctx: &ExtensionContext<'_>, err: &ServerError) {
         if let Some(parent_cx) = self.contexts.get(&EXECUTE_CTX).cloned() {
-            parent_cx.span().add_event(
-                "error".to_string(),
-                vec![KeyValue::new(KEY_ERROR, err.to_string())],
-            );
+            parent_cx
+                .span()
+                .add_event("error".to_string(), vec![KEY_ERROR.string(err.to_string())]);
         }
     }
 }