+++ src/resolver_utils/container.rs
@@ -2,7 +2,7 @@ use std::collections::BTreeMap;
 use std::future::Future;
 use std::pin::Pin;
 
-use crate::extensions::{ErrorLogger, ExtensionContext, ResolveInfo};
+use crate::extensions::{ErrorLogger, ResolveInfo};
 use crate::parser::types::Selection;
 use crate::registry::MetaType;
 use crate::{
@@ -183,11 +183,6 @@ impl<'a> Fields<'a> {
                                     }
                                 }?
                             } else {
-                                let ctx_extension = ExtensionContext {
-                                    schema_data: &ctx.schema_env.data,
-                                    query_data: &ctx.query_env.ctx_data,
-                                };
-
                                 let type_name = T::type_name();
                                 let resolve_info = ResolveInfo {
                                     resolve_id: ctx_field.resolve_id,
@@ -215,10 +210,7 @@ impl<'a> Fields<'a> {
                                     },
                                 };
 
-                                ctx_field
-                                    .query_env
-                                    .extensions
-                                    .resolve_start(&ctx_extension, &resolve_info);
+                                ctx_field.query_env.extensions.resolve_start(&resolve_info);
 
                                 let res = match root.resolve_field(&ctx_field).await {
                                     Ok(value) => Ok((field_name, value.unwrap_or_default())),
@@ -226,13 +218,8 @@ impl<'a> Fields<'a> {
                                         Err(e.path(PathSegment::Field(field_name.to_string())))
                                     }
                                 }
-                                .log_error(&ctx_extension, &ctx_field.query_env.extensions)?;
-
-                                ctx_field
-                                    .query_env
-                                    .extensions
-                                    .resolve_end(&ctx_extension, &resolve_info);
-
+                                .log_error(&ctx_field.query_env.extensions)?;
+                                ctx_field.query_env.extensions.resolve_end(&resolve_info);
                                 res
                             };