+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.7.0]
+
+## Fixed
+
+- Fix chrono-tz integration. [#452](https://github.com/async-graphql/async-graphql/pull/452)
+
+## Changed
+
+- Rework Extension & TracingExtension & OpenTelemetryExtension
+
+## [2.6.5] - 2021-03-24
+
+- In websocket, if the client sends `start` before `connection_init`, the connection will be immediately disconnected and return `1011` error. [#451](https://github.com/async-graphql/async-graphql/issues/451)
+
 ## [2.6.4] - 2021-03-22
 
 - Fix docs.